update README with examples for DQDL and add the list of supported DQDL rules
